practice_applications/text_data_pipeline/requirements.txt
python-dotenv
requests
confluent-kafka
apache-airflow
newsapi-python
psycopg2
sqlalchemy
